Move the lastVerifiedBatch polling to the activity view
